Add the /post_update url, which updates master
Summary:
Add a url which github can ping when we make changes to our master
branch, so that it will keep the Khan:master branch up to date, which is
important because we link all of the issues on github to it for testing.
Internally, it just calls the same thing that requesting /branch/Khan:master
does.

Test Plan: Go to /post_update, see it update master

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2080